requirements.txt
Django==1.9.5
djangorestframework==3.3.3
MySQL-python
django-filter==0.13.0
